Rip from vitis-integration-ontop-fsim-mill
