+++ deploy/runtools/firesim_topology_elements.py
@@ -281,18 +281,20 @@ class FireSimServerNode(FireSimNode):
             rootLogger.debug("[localhost] " + str(localcap))
             rootLogger.debug("[localhost] " + str(localcap.stderr))
 
+        dest_sim_dir = self.host_instance.dest_simulation_dir
+
         # mount rootfs, copy files from it back to local system
         rfsname = self.get_rootfs_name()
         if rfsname is not None:
             is_qcow2 = rfsname.endswith(".qcow2")
-            mountpoint = """/home/centos/sim_slot_{}/mountpoint""".format(simserverindex)
+            mountpoint = """{}/sim_slot_{}/mountpoint""".format(dest_sim_dir, simserverindex)
             with StreamLogger('stdout'), StreamLogger('stderr'):
                 run("""sudo mkdir -p {}""".format(mountpoint))
 
                 if is_qcow2:
                     rfsname = self.get_host_instance().nbd_tracker.get_nbd_for_imagename(rfsname)
                 else:
-                    rfsname = """/home/centos/sim_slot_{}/{}""".format(simserverindex, rfsname)
+                    rfsname = """{}/sim_slot_{}/{}""".format(dest_sim_dir, simserverindex, rfsname)
 
                 run("""sudo mount {blockfile} {mntpt}""".format(blockfile=rfsname, mntpt=mountpoint))
                 with warn_only():
@@ -324,7 +326,7 @@ class FireSimServerNode(FireSimNode):
 
         ## copy output files generated by the simulator that live on the host:
         ## e.g. uartlog, memory_stats.csv, etc
-        remote_sim_run_dir = """/home/centos/sim_slot_{}/""".format(simserverindex)
+        remote_sim_run_dir = """{}/sim_slot_{}/""".format(dest_sim_dir, simserverindex)
         for simoutputfile in jobinfo.simoutputs:
             with warn_only(), StreamLogger('stdout'), StreamLogger('stderr'):
                 rsync_cap = rsync_project(remote_dir=remote_sim_run_dir + simoutputfile,
@@ -612,9 +614,11 @@ class FireSimSwitchNode(FireSimNode):
             rootLogger.debug("[localhost] " + str(localcap))
             rootLogger.debug("[localhost] " + str(localcap.stderr))
 
+        dest_sim_dir = self.host_instance.dest_simulation_dir
+
         ## copy output files generated by the simulator that live on the host:
         ## e.g. uartlog, memory_stats.csv, etc
-        remote_sim_run_dir = """/home/centos/switch_slot_{}/""".format(switch_slot_no)
+        remote_sim_run_dir = """{}/switch_slot_{}/""".format(dest_sim_dir, switch_slot_no)
         for simoutputfile in ["switchlog"]:
             with StreamLogger('stdout'), StreamLogger('stderr'):
                 get(remote_path=remote_sim_run_dir + simoutputfile, local_path=job_dir)
